@@ -2,86 +2,72 @@
   "name": "ketaka-lite",
   "version": "0.0.0",
   "description": "",
-  "main": "main.js",
-  "jest": {
-    "scriptPreprocessor": "<rootDir>/preprocessor.js",
-    "testFileExtensions": [
-      "js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json"
-    ],
-    "globals": {
-      "SCRIPTS": "<rootDir>/src/scripts"
-    },
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "<rootDir>/node_modules/"
-    ],
-    "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/react"
-    ]
-  },
+  "main": "index.js",
   "dependencies": {
-    "animate.css": "^3.4.0",
-    "archiver": "^0.15.1",
+    "animate.css": "3.5.1",
+    "archiver": "0.15.1",
     "bootstrap-sass": "3.3.5",
-    "bufferhelper": "^0.2.0",
+    "bufferhelper": "0.2.0",
     "check-tibetan": "https://github.com/karmapa/checkTibetan_for_ketaka.git#e4b3758",
-    "classnames": "^2.1.3",
-    "codemirror": "^5.4.0",
-    "csv": "^0.4.6",
+    "classnames": "2.1.3",
+    "codemirror": "5.12.0",
+    "csv": "0.4.6",
     "decompress": "3.0.0",
-    "event-stream": "^3.3.1",
-    "eventemitter3": "^1.1.1",
-    "file-type": "^2.10.2",
+    "event-stream": "3.3.1",
+    "eventemitter3": "1.1.1",
+    "file-type": "2.10.2",
     "fs-extra": "0.26.2",
-    "htmlparser": "^1.7.7",
-    "humps": "^0.6.0",
+    "htmlparser": "1.7.7",
+    "humps": "0.6.0",
     "javascript-natural-sort": "0.7.1",
     "keycode": "2.1.0",
-    "keypress.js": "^2.1.0",
-    "lodash": "^3.10.0",
-    "mkdirp": "^0.5.1",
-    "node-uuid": "^1.4.3",
-    "react": "0.13.3",
-    "react-bootstrap": "0.23.7",
-    "react-codemirror": "https://github.com/kmsheng/react-codemirror.git#fix-performance-hit",
-    "react-list": "0.6.3",
+    "keypress.js": "2.1.0",
+    "lodash": "3.10.0",
+    "mkdirp": "0.5.1",
+    "node-uuid": "1.4.3",
+    "path": "0.12.7",
+    "react": "0.14.7",
+    "react-bootstrap": "0.28.3",
+    "react-codemirror": "0.2.5",
+    "react-dom": "0.14.7",
     "react-pure-render": "1.0.2",
-    "react-redux": "^0.2.2",
+    "react-redux": "4.4.0",
     "react-router": "1.0.0-beta2",
-    "react-toastr": "^1.5.1",
-    "read-chunk": "^1.0.1",
-    "redux": "1.0.0-rc",
-    "redux-thunk": "^0.1.0",
-    "rimraf": "^2.4.3",
-    "through2": "^2.0.0",
-    "toastr": "^2.1.2",
-    "whatwg-fetch": "0.9.0",
-    "zpad": "^0.5.0"
+    "react-toastr": "2.4.0",
+    "read-chunk": "1.0.1",
+    "redux": "3.3.1",
+    "redux-thunk": "1.0.3",
+    "rimraf": "2.4.3",
+    "through2": "2.0.0",
+    "toastr": "2.1.2",
+    "zpad": "0.5.0"
   },
   "devDependencies": {
-    "ava": "^0.4.2",
-    "babel": "5.8.21",
-    "babel-core": "5.8.22",
+    "ava": "0.12.0",
+    "babel": "6.5.2",
+    "babel-core": "6.5.2",
     "babel-eslint": "5.0.0",
-    "babel-jest": "5.3.0",
-    "babel-loader": "5.3.2",
+    "babel-loader": "6.2.3",
+    "babel-plugin-transform-async-to-generator": "6.5.0",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
+    "babel-polyfill": "6.5.0",
+    "babel-preset-es2015": "6.5.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-stage-0": "6.5.0",
+    "babel-register": "6.5.2",
     "babel-runtime": "5.8.20",
-    "concurrently": "^1.0.0",
+    "concurrently": "1.0.0",
     "css-loader": "0.15.4",
-    "electron-prebuilt": "0.33.6",
-    "eslint": "1.1.0",
+    "electron-prebuilt": "0.36.8",
+    "eslint": "2.2.0",
     "eslint-loader": "1.0.0",
-    "eslint-plugin-react": "3.2.2",
+    "eslint-plugin-react": "4.0.0",
     "file-loader": "0.8.4",
     "html-loader": "0.3.0",
     "imports-loader": "0.6.4",
-    "jest-cli": "0.8.0",
     "node-libs-browser": "0.5.3",
     "node-sass": "3.4.2",
+    "nodemon": "1.9.0",
     "object-assign": "2.0.0",
     "react-hot-loader": "1.2.8",
     "react-tools": "0.13.3",
