@@ -1,77 +1,70 @@
-var BrowserWindow = require('browser-window');
-var app = require('app');
-var ipc = require('ipc');
-var ipcHandlers = require('./main/ipcHandlers');
-var bindEventName = require('./main/decorators/bindEventName');
-var Helper = require('./main/services/Helper');
-var MenuConfig = require('./main/services/MenuConfig');
-var PATH_APP_DOC = require('./main/constants').PATH_APP_DOC;
-var Menu = require('menu');
+import {app, ipcMain, crashReporter, screen, BrowserWindow, Menu} from 'electron';
+import * as controllers from './main/controllers';
+import bindEventName from './main/decorators/bindEventName';
+import {Helper, MenuConfig} from './main/services';
+import {PATH_APP_DOC} from './main/constants/appConstants';
 
-require('crash-reporter').start();
+crashReporter.start({
+  companyName: 'dharma-treasure',
+  submitURL: 'https://log.dharma-treasure.org/ketaka-lite'
+});
 
-var mainWindow = null;
+let mainWindow = null;
 
-app.on('window-all-closed', function() {
+app.on('window-all-closed', () => {
   app.quit();
 });
 
-app.on('ready', function() {
+app.on('ready', () => {
 
-  var screen = require('screen');
-  var size = screen.getPrimaryDisplay().workAreaSize;
-
+  let size = screen.getPrimaryDisplay().workAreaSize;
   mainWindow = new BrowserWindow({width: size.width, height: size.height});
 
   Helper.mkdirp(PATH_APP_DOC)
-    .then(function() {
+    .then(() => {
 
-      mainWindow.loadUrl('file://' + __dirname + '/index.html');
+      mainWindow.loadURL('file://' + __dirname + '/index.html');
 
-      mainWindow.on('closed', function() {
+      mainWindow.on('closed', () => {
         mainWindow = null;
       });
     });
 });
 
-app.once('ready', function() {
-
-  if (Menu.getApplicationMenu()) {
- //   return;
-  }
-  var template = MenuConfig.getTemplate({mainWindow: mainWindow});
-  var menu = Menu.buildFromTemplate(template);
+app.once('ready', () => {
+  let template = MenuConfig.getTemplate({mainWindow: mainWindow});
+  let menu = Menu.buildFromTemplate(template);
   Menu.setApplicationMenu(menu);
 });
 
-ipc = bindEventName(ipc);
+const ipc = bindEventName(ipcMain);
 
-ipc.on('import-button-clicked', ipcHandlers.importButtonClicked);
+ipc.on('import-button-clicked', controllers.importButtonClicked);
 
-ipc.on('import-zip', ipcHandlers.importZip);
+ipc.on('import-zip', controllers.importZip);
 
-ipc.on('save', ipcHandlers.save);
+ipc.on('save', controllers.save);
 
-ipc.on('save-as', ipcHandlers.saveAs);
+ipc.on('save-as', controllers.saveAs);
 
-ipc.on('page-image-upload-button-clicked', ipcHandlers.pageImageUploadButtonClicked);
+ipc.on('page-image-upload-button-clicked', controllers.pageImageUploadButtonClicked);
 
-ipc.on('add-doc', ipcHandlers.addDoc);
+ipc.on('add-doc', controllers.addDoc);
 
-ipc.on('find-doc-names', ipcHandlers.findDocNames);
+ipc.on('find-doc-names', controllers.findDocNames);
 
-ipc.on('change-doc-settings', ipcHandlers.changeDocSettings);
+ipc.on('change-doc-settings', controllers.changeDocSettings);
 
-ipc.on('open', ipcHandlers.open);
+ipc.on('open', controllers.open);
 
-ipc.on('open-bamboo', ipcHandlers.openBamboo);
+ipc.on('open-bamboo', controllers.openBamboo);
 
-ipc.on('delete-doc', ipcHandlers.deleteDoc);
+ipc.on('delete-doc', controllers.deleteDoc);
 
-ipc.on('export-zip', ipcHandlers.exportZip);
+ipc.on('export-zip', controllers.exportZip);
 
-ipc.on('export-file-with-pb', ipcHandlers.exportFileWithPb);
+ipc.on('export-file-with-pb', controllers.exportFileWithPb);
 
-ipc.on('add-pb-files', ipcHandlers.addPbFiles);
+ipc.on('add-pb-files', controllers.addPbFiles);
 
-ipc.on('get-app-data', ipcHandlers.getAppData);
+ipc.on('get-app-data', controllers.getAppData);
