@@ -1,9 +1,4 @@
-require('babel/polyfill');
-require('./index.scss');
-
-import React from 'react';
-import HashHistory from 'react-router/lib/HashHistory';
-import Root from './root';
-
-const history = new HashHistory();
-React.render(<Root history={history} />, document.getElementById('root'));
+// install babel hooks in the main process
+require('babel-register');
+require('babel-polyfill');
+require('./main.js');
